Added option to get data to be used for creating charts. Added classes to handle chart requests and responses.

+++ src/main/java/com/semafoor/semaforce/model/dto/results/WeeklyResultDto.java
@@ -16,9 +16,9 @@ public class WeeklyResultDto {
     private int exerciseNumber;
     private int weekNumber;
 
-    private List<Integer> weightsLifted;
+    private List<Double> weightsLifted;
     private List<Integer> repetitionsPerformed;
-    private List<Integer> rpe;
+    private List<Double> rpe;
 
     private Long exerciseId;
